requirements.txt
cached-property==1.4.3
certifi==2018.4.16
chardet==3.0.4
click==6.7
diff-match-patch==20121119
Django==2.0.5
django-admin-json-editor==0.1.5
django-ckeditor==5.5.0
django-environ==0.4.4
django-extensions==2.0.7
django-import-export==1.0.1
django-js-asset==1.1.0
django-json-widget==0.1.1
django-modeladmin-reorder==0.3.1
django-reversion==2.0.13
django-reversion-compare==0.8.4
docker==3.3.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
edtf==4.0.1
et-xmlfile==1.0.1
Faker==0.8.17
freeze==1.0.10
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.8.1
idna==2.6
jdcal==1.4
jsonschema==2.6.0
numpy==1.14.5
odfpy==1.3.6
openpyxl==2.5.4
pathlib==1.0.1
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.5
pygdal==2.1.2.3
pyparsing==2.2.0
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.3
PyYAML==3.12
raven==6.5.0
requests==2.18.4
six==1.11.0
tablib==0.12.1
text-unidecode==1.2
texttable==0.9.1
unicodecsv==0.14.1
urllib3==1.22
websocket-client==0.48.0
Werkzeug==0.14.1
Wheelhouse==0.1.3
xlrd==1.1.0
xlwt==1.3.0
